+++ src/modules/PlayerBots/playerbot/strategy/triggers/TriggerContext.h
@@ -344,6 +344,7 @@ namespace ai
             creators["chromaggus affliction danger"] = [](PlayerbotAI* ai) { return new ChromaggusAfflictionDangerTrigger(ai); };
             creators["start firemaw fight"] = [](PlayerbotAI* ai) { return new FiremawStartFightTrigger(ai); };
             creators["end firemaw fight"] = [](PlayerbotAI* ai) { return new FiremawEndFightTrigger(ai); };
+            creators["firemaw flame buffet swap"] = [](PlayerbotAI* ai) { return new FiremawFlameBuffetSwapTrigger(ai); };
             creators["start ebonroc fight"] = [](PlayerbotAI* ai) { return new EbonrocStartFightTrigger(ai); };
             creators["end ebonroc fight"] = [](PlayerbotAI* ai) { return new EbonrocEndFightTrigger(ai); };
             creators["start flamegor fight"] = [](PlayerbotAI* ai) { return new FlamegorStartFightTrigger(ai); };